adaptation
git-svn-id: http://svn.miranda-ng.org/main/trunk@1964 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/!NotAdopted/Skype/debug.c
@@ -45,7 +45,7 @@ void do_log(const char *pszFormat, ...) {
 	char *ct, *pNewBuf;
 	va_list ap;
 	time_t lt;
-	int iLen;
+	INT64 iLen;
 
 	if (!m_szLogBuf || !m_fpLogFile) return;
 	EnterCriticalSection(&m_WriteFileMutex);
